pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iops-benchmark"
dynamic = ["version"]
description = "A generic benchmark orchestration framework for automated parametric experiments"
readme = "PYPI_README.md"
requires-python = ">=3.10"
license = {text="BSD-3-Clause"}
authors = [
    {name = "TADAAM - INRIA Bordeaux", email = "luan.teylo@inria.fr"}
]
keywords = ["benchmark", "performance", "testing", "parametric", "sweep", "hpc", "slurm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Benchmark",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
]
dependencies = [
    "pyyaml",
    "scikit-optimize",
    "pandas>=2.1.0",
    "jinja2>=3.0.0",
    "plotly",
]

[project.optional-dependencies]
parquet = ["pyarrow"]
watch = ["rich"]
plots = ["kaleido"]

[project.urls]
Homepage = "https://gitlab.inria.fr/lgouveia/iops"
Documentation = "https://gitlab.inria.fr/lgouveia/iops"
Repository = "https://gitlab.inria.fr/lgouveia/iops"
Issues = "https://gitlab.inria.fr/lgouveia/iops/-/issues"

[project.scripts]
iops = "iops.main:main"

[tool.setuptools.packages.find]
include = ["iops*", "docs*"]

[tool.setuptools.package-data]
iops = ["VERSION", "**/*.yaml", "**/*.yml"]
docs = ["**/*.md", "**/*.yaml", "**/*.yml", "**/*.py", "**/*.sh", "**/*.txt"]

[tool.setuptools.dynamic]
version = {file = ["iops/VERSION"]}
